+++ projects/ngx-smooth-scroll-demo/src/app/app.component.ts
@@ -3,6 +3,8 @@ import { Component, ViewChild, ElementRef } from '@angular/core';
 
 /** Custom Modules */
 import { NgxSmoothScrollService, NgxSmoothScrollOption } from 'projects/ngx-smooth-scroll/src/public-api';
+import { NgxSmoothScrollKeyboardDirective } from 'projects/ngx-smooth-scroll/src/lib/ngx-smooth-scroll-keyboard.directive';
+import { NgxSmoothScrollWheelDirective } from 'projects/ngx-smooth-scroll/src/lib/ngx-smooth-scroll-wheel.directive';
 
 interface ScrollVariables {
   enabled: boolean;
@@ -37,7 +39,8 @@ export class AppComponent {
       timingFunction: 'ease'
     },
     direction: 'vertical',
-    skip: 0
+    skip: 0,
+    animating: false
   }
 
   public keyboard: ScrollVariables = {
@@ -47,10 +50,13 @@ export class AppComponent {
       timingFunction: 'ease'
     },
     direction: 'vertical',
-    skip: 0
+    skip: 0,
+    animating: false
   }
 
   @ViewChild('basicScroll', { static: false }) basicScrollElRef: ElementRef;
+  @ViewChild(NgxSmoothScrollKeyboardDirective, { static: false }) keyboardDirectiveRef: NgxSmoothScrollKeyboardDirective;
+  @ViewChild(NgxSmoothScrollWheelDirective, { static: false }) wheelDirectiveRef: NgxSmoothScrollWheelDirective;
 
   constructor(
     private smoothScroll: NgxSmoothScrollService
@@ -64,6 +70,21 @@ export class AppComponent {
 
   public durationLabel(duration: number): string {
     return `${ Math.floor(duration / 100) / 10 }s`;
+  }
+
+  public animationState(type: string, value: boolean): void {
+    this[type].animating = value;
+  }
+
+  public interrupt(type: string): void {
+    switch (type) {
+      case 'keyboard':
+        this.keyboardDirectiveRef.interrupt();
+        break;
+      case 'wheel':
+        this.wheelDirectiveRef.interrupt();
+        break;
+    }
   }
 
 }