+++ projects/ngx-smooth-scroll/CHANGELOG.md
@@ -1,2 +1,8 @@
-### 0.1.2 Update README
-### 0.1.1 Update README
+### 0.2.0-beta Add interrupt method to directives & Fix bug
+- `NgxSmoothScrollWheelDirective.interrupt`: () => boolean
+    - **@return**: boolean, whether behavior interrupted successfully.
+- `NgxSmoothScrollKeyboardDirective.interrupt`: () => boolean
+    - **@return**: boolean, whether behavior interrupted successfully.
+- Fix bug: `NgxSmoothScroll.interrupt` not worked propery.
+### 0.1.2-beta Update README
+### 0.1.1-beta Update README